file tidy up and style sheet imports
