+++ src/Screens/UserSignUp.js
@@ -1,0 +1,16 @@
+import { StatusBar } from 'expo-status-bar';
+import { StyleSheet, Text, Button, View } from 'react-native';
+import { useNavigation } from '@react-navigation/native';
+import { useDrawerStatus } from '@react-navigation/drawer';
+import styles from '../components/ReusableStyleSheet';
+
+export default function MeditationTracker() {
+    const nav = useNavigation();
+    const drawerStatus = useDrawerStatus();
+    console.log(drawerStatus);
+    return (
+        <View style={styles.container}>
+            <StatusBar style="auto" />
+    </View>
+);
+}
