+++ src/components/ReusableStyleSheet.js
@@ -1,14 +1,16 @@
-import {StyleSheet} from 'react-native';
+import { StyleSheet } from 'react-native';
 
 // add things here as needed that can be applied to all screens
 const styles = StyleSheet.create({
     container: {
-    flex: 1,
-    backgroundColor: '#15224c',
-    alignItems: 'center',
-    justifyContent: 'center',
+        flex: 1,
+        backgroundColor: '#15224c',
+        alignItems: 'center',
+        justifyContent: 'center',
     },
     text: {
         color: '#e0ab78'
     }
 });
+
+export default styles;
